Fixed maingopy import
